+++ docs/app/supported-games/[appName]/AdComponent.tsx
@@ -28,7 +28,7 @@ export default function AdComponent() {
           setAdKey(prev => prev + 1)
           scheduleNextRefresh()
         },
-        2 * 60 * 1000,
+        1.7 * 60 * 1000,
       )
 
       return interval